+++ src/adapter/src/coord/sequencer/inner.rs
@@ -65,6 +65,7 @@ use mz_sql::plan::{
 };
 use mz_sql::pure::{PurifiedSourceExport, generate_subsource_statements};
 use mz_storage_types::sinks::StorageSinkDesc;
+use mz_timestamp_oracle::TimestampOracle;
 // Import `plan` module, but only import select elements to avoid merge conflicts on use statements.
 use mz_sql::plan::{
     AlterConnectionAction, AlterConnectionPlan, CreateSourcePlanBundle, ExplainSinkSchemaPlan,
@@ -146,6 +147,30 @@ macro_rules! return_if_err {
 }
 
 pub(super) use return_if_err;
+
+fn spawn_linearized_read_ts<S>(
+    oracle: Option<Arc<dyn TimestampOracle<Timestamp> + Send + Sync>>,
+    name: &'static str,
+    build_stage: impl FnOnce(Option<Timestamp>) -> S + Send + 'static,
+) -> StageResult<Box<S>>
+where
+    S: Send + 'static,
+{
+    match oracle {
+        Some(oracle) => {
+            let span = Span::current();
+            StageResult::Handle(mz_ore::task::spawn(
+                move || name,
+                async move {
+                    let oracle_read_ts = oracle.read_ts().await;
+                    Ok(Box::new(build_stage(Some(oracle_read_ts))))
+                }
+                .instrument(span),
+            ))
+        }
+        None => StageResult::Immediate(Box::new(build_stage(None))),
+    }
+}
 
 struct DropOps {
     ops: Vec<catalog::Op>,
